+++ wand/api.py
@@ -218,7 +218,7 @@ try:
     library.MagickWriteImageFile.argtypes = [ctypes.c_void_p, ctypes.c_void_p]
 
     library.MagickGetImageResolution.argtypes = [
-        ctypes.c_void_p, ctypes.POINTER(ctypes.c_double),
+        ctypes.c_void_p,
         ctypes.POINTER(ctypes.c_double)
     ]
 
@@ -228,7 +228,6 @@ try:
 
     library.MagickSetResolution.argtypes = [ctypes.c_void_p, ctypes.c_double,
                                             ctypes.c_double]
-    library.MagickSetResolution.restype = ctypes.c_bool
 
     library.MagickGetImageWidth.argtypes = [ctypes.c_void_p]
     library.MagickGetImageWidth.restype = ctypes.c_size_t
@@ -372,12 +371,195 @@ try:
 
     library.MagickTrimImage.argtypes = [ctypes.c_void_p]
 
+    library.MagickGetSize.argtypes = [ctypes.c_void_p,
+                                      ctypes.POINTER(ctypes.c_uint),
+                                      ctypes.POINTER(ctypes.c_uint)]
+    library.MagickGetSize.restype = ctypes.c_int
+                                      
+    library.MagickSetSize.argtypes = [ctypes.c_void_p,
+                                      ctypes.c_uint,
+                                      ctypes.c_uint]
+    library.MagickSetSize.restype = ctypes.c_int
+
+    library.MagickGetFont.argtypes = [ctypes.c_void_p]
+    library.MagickGetFont.restype = ctypes.c_char_p
+
+    library.MagickSetFont.argtypes = [ctypes.c_void_p,
+                                      ctypes.c_char_p]
+    library.MagickSetFont.restype = ctypes.c_int
+
+    library.MagickGetPointsize.argtypes = [ctypes.c_void_p]
+    library.MagickGetPointsize.restype = ctypes.c_double
+
+    library.MagickSetPointsize.argtypes = [ctypes.c_void_p,
+                                           ctypes.c_double]
+    library.MagickSetPointsize.restype = ctypes.c_int
+
+    library.MagickGetGravity.argtypes = [ctypes.c_void_p]
+    library.MagickGetGravity.restype = ctypes.c_int
+
+    library.MagickSetGravity.argtypes = [ctypes.c_void_p,
+                                         ctypes.c_int]
+    library.MagickSetGravity.restype = ctypes.c_int
+
+    library.MagickSetLastIterator.argtypes = [ctypes.c_void_p]
+
+    library.MagickGetBackgroundColor.argtypes = [ctypes.c_void_p]
+    library.MagickGetBackgroundColor.restype = ctypes.c_void_p
+
+    library.MagickSetBackgroundColor.argtypes = [ctypes.c_void_p,
+                                                 ctypes.c_void_p]
+    library.MagickSetBackgroundColor.restype = ctypes.c_int
+
+    library.MagickGetOption.argtypes = [ctypes.c_void_p,
+                                        ctypes.c_char_p]
+    library.MagickGetOption.restype = ctypes.c_char_p
+
+    library.MagickSetOption.argtypes = [ctypes.c_void_p,
+                                        ctypes.c_char_p,
+                                        ctypes.c_char_p]
+    library.MagickSetOption.restype = ctypes.c_int
+
+    library.MagickGetAntialias.argtypes = [ctypes.c_void_p]
+    library.MagickGetAntialias.restype = ctypes.c_int
+
+    library.MagickSetAntialias.argtypes = [ctypes.c_void_p,
+                                           ctypes.c_int]
+    library.MagickSetAntialias.restype = ctypes.c_int
+
     # These functions are const so it's okay for them to be c_char_p
     libmagick.GetMagickVersion.argtypes = [ctypes.POINTER(ctypes.c_size_t)]
     libmagick.GetMagickVersion.restype = ctypes.c_char_p
 
     libmagick.GetMagickReleaseDate.argtypes = []
     libmagick.GetMagickReleaseDate.restype = ctypes.c_char_p
+
+    library.NewDrawingWand.restype = ctypes.c_void_p
+
+    library.CloneDrawingWand.argtypes = [ctypes.c_void_p]
+    library.CloneDrawingWand.restype = ctypes.c_void_p
+
+    library.DestroyDrawingWand.argtypes = [ctypes.c_void_p]
+    library.DestroyDrawingWand.restype = ctypes.c_void_p
+
+    library.IsDrawingWand.argtypes = [ctypes.c_void_p]
+    library.IsDrawingWand.restype = ctypes.c_int
+
+    library.DrawGetException.argtypes = [ctypes.c_void_p,
+                                         ctypes.POINTER(ctypes.c_int)]
+    library.DrawGetException.restype = ctypes.c_char_p
+
+    library.DrawClearException.argtypes = [ctypes.c_void_p]
+    library.DrawClearException.restype = ctypes.c_int
+
+    library.DrawSetFont.argtypes = [ctypes.c_void_p,
+                                    ctypes.c_char_p]
+
+    library.DrawSetFontSize.argtypes = [ctypes.c_void_p,
+                                        ctypes.c_double]
+
+    library.DrawSetFillColor.argtypes = [ctypes.c_void_p,
+                                         ctypes.c_void_p]
+
+    library.DrawSetTextAlignment.argtypes = [ctypes.c_void_p,
+                                             ctypes.c_int]
+
+    library.DrawSetTextAntialias.argtypes = [ctypes.c_void_p,
+                                             ctypes.c_int]
+
+    library.DrawSetTextDecoration.argtypes = [ctypes.c_void_p,
+                                              ctypes.c_int]
+
+    library.DrawSetTextEncoding.argtypes = [ctypes.c_void_p,
+                                            ctypes.c_char_p]
+
+    library.DrawSetTextInterlineSpacing.argtypes = [ctypes.c_void_p,
+                                                    ctypes.c_double]
+
+    library.DrawSetTextInterwordSpacing.argtypes = [ctypes.c_void_p,
+                                                    ctypes.c_double]
+
+    library.DrawSetTextKerning.argtypes = [ctypes.c_void_p,
+                                           ctypes.c_double]
+
+    library.DrawSetTextUnderColor.argtypes = [ctypes.c_void_p,
+                                              ctypes.c_void_p]
+
+    library.DrawGetFillColor.argtypes = [ctypes.c_void_p,
+                                         ctypes.c_void_p]
+
+    library.DrawGetFont.argtypes = [ctypes.c_void_p]
+    library.DrawGetFont.restype = ctypes.c_char_p
+
+    library.DrawGetFontSize.argtypes = [ctypes.c_void_p]
+    library.DrawGetFontSize.restype = ctypes.c_double
+
+    library.DrawGetTextAlignment.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextAlignment.restype = ctypes.c_int
+
+    library.DrawGetTextAntialias.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextAntialias.restype = ctypes.c_int
+
+    library.DrawGetTextDecoration.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextDecoration.restype = ctypes.c_int
+
+    library.DrawGetTextEncoding.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextEncoding.restype = ctypes.c_char_p
+
+    library.DrawGetTextInterlineSpacing.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextInterlineSpacing.restype = ctypes.c_double
+
+    library.DrawGetTextInterwordSpacing.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextInterwordSpacing.restype = ctypes.c_double
+
+    library.DrawGetTextKerning.argtypes = [ctypes.c_void_p]
+    library.DrawGetTextKerning.restype = ctypes.c_double
+
+    library.DrawGetTextUnderColor.argtypes = [ctypes.c_void_p,
+                                              ctypes.c_void_p]
+
+    library.DrawSetGravity.argtypes = [ctypes.c_void_p,
+                                       ctypes.c_int]
+
+    library.DrawGetGravity.argtypes = [ctypes.c_void_p]
+    library.DrawGetGravity.restype = ctypes.c_int
+
+    library.MagickAnnotateImage.argtypes = [ctypes.c_void_p,
+                                            ctypes.c_void_p,
+                                            ctypes.c_double,
+                                            ctypes.c_double,
+                                            ctypes.c_double,
+                                            ctypes.c_char_p]
+    library.MagickAnnotateImage.restype = ctypes.c_int
+
+    library.ClearDrawingWand.argtypes = [ctypes.c_void_p]
+
+    library.MagickDrawImage.argtypes = [ctypes.c_void_p,
+                                        ctypes.c_void_p]
+    library.MagickDrawImage.restype = ctypes.c_int
+
+    library.DrawLine.argtypes = [ctypes.c_void_p,
+                                 ctypes.c_double,
+                                 ctypes.c_double,
+                                 ctypes.c_double,
+                                 ctypes.c_double]
+
+    library.DrawAnnotation.argtypes = [ctypes.c_void_p,
+                                       ctypes.c_double,
+                                       ctypes.c_double,
+                                       ctypes.POINTER(ctypes.c_ubyte)]
+
+    library.MagickQueryFontMetrics.argtypes = [ctypes.c_void_p,
+                                               ctypes.c_void_p,
+                                               ctypes.c_char_p]
+    library.MagickQueryFontMetrics.restype = ctypes.POINTER(ctypes.c_double)
+
+    library.MagickQueryMultilineFontMetrics.argtypes = [ctypes.c_void_p,
+                                                        ctypes.c_void_p,
+                                                        ctypes.c_char_p]
+    library.MagickQueryMultilineFontMetrics.restype = ctypes.POINTER(
+        ctypes.c_double
+    )
 except AttributeError:
     raise ImportError('MagickWand shared library not found or incompatible')