+++ wandtests/image.py
@@ -15,6 +15,7 @@ from wand.version import MAGICK_VERSION_INFO
 from wand.image import ClosedImageError, Image
 from wand.color import Color
 from wand.exceptions import MissingDelegateError
+from wand.font import Font
 
 
 skip_slow_tests = bool(os.environ.get('WANDTESTS_SKIP_SLOW_TESTS'))
@@ -45,7 +46,6 @@ tests = Tests()
 
 def asset(filename):
     return os.path.join(os.path.dirname(__file__), 'assets', filename)
-
 
 @tests.test
 def empty_image():
@@ -1105,3 +1105,45 @@ def border():
             assert img[2, -6] == left_bottom
             assert img[-3, 5] == right_top
             assert img[-3, -6] == right_bottom
+
+
+@tests.test
+def caption():
+    with Image(width=144, height=192, background=Color('#1e50a2')) as img:
+        font = Font(
+            path=asset('League_Gothic.otf'),
+            color=Color("gold"),
+            size=12,
+            antialias=False
+        )
+        img.caption(
+            'Test message',
+            font=font,
+            left=5, top=144,
+            width=134, height=20,
+            gravity='center'
+        )
+
+
+@tests.test
+def setfont():
+    with Image(width=144, height=192, background=Color('#1e50a2')) as img:
+        font = Font(
+            path=asset('League_Gothic.otf'),
+            color=Color('gold'),
+            size=12,
+            antialias=False
+        )
+        img.font = font
+        assert img.font_path == font.path
+        assert img.font_size == font.size
+        assert img.font_color == font.color
+        assert img.font_antialias == font.antialias
+        assert img.font == font
+
+
+@tests.test
+def setgravity():
+    with Image(width=144, height=192, background=Color('#1e50a2')) as img:
+        img.gravity = 'center'
+        assert img.gravity == 'center'